+++ cdef/time/time.h
@@ -32,3 +32,29 @@ char *strptime(const char *s, const char *format, struct tm *tm);
 size_t strftime(char *restrict s, size_t maxsize, const char *restrict format, const struct tm *restrict timeptr);
 time_t timegm(struct tm *timeptr);
 time_t mktime(struct tm *timeptr);
+
+
+struct rusage {
+	struct timeval ru_utime; /* user time used */
+	struct timeval ru_stime; /* system time used */
+	long ru_maxrss;          /* max resident set size */
+	long ru_ixrss;           /* integral shared text memory size */
+	long ru_idrss;           /* integral unshared data size */
+	long ru_isrss;           /* integral unshared stack size */
+	long ru_minflt;          /* page reclaims */
+	long ru_majflt;          /* page faults */
+	long ru_nswap;           /* swaps */
+	long ru_inblock;         /* block input operations */
+	long ru_oublock;         /* block output operations */
+	long ru_msgsnd;          /* messages sent */
+	long ru_msgrcv;          /* messages received */
+	long ru_nsignals;        /* signals received */
+	long ru_nvcsw;           /* voluntary context switches */
+	long ru_nivcsw;          /* involuntary context switches */
+};
+
+static const int RUSAGE_SELF     =   0;
+static const int RUSAGE_CHILDREN =  -1;
+
+int
+getrusage(int who, struct rusage *r_usage);